SplitGraph::RemoveTrivialSplits no longer relies on mixing mixing an iterator with pop_back() (as that doesn't work reliably on Windows) (besides, it was relying on undefined behaviour!).  It now returns size_t rather than int.

+++ pda/splitgraph.h
@@ -382,7 +382,7 @@ public:
 	 * remove all trivial splits
 	 * @return number of trivial splits removed
 	*/
-	int removeTrivialSplits();
+	size_t removeTrivialSplits();
     
 protected:
